extracted wrapper for Rule to a dedicated class
