+++ diktat-rules/src/main/kotlin/org/cqfn/diktat/ruleset/rules/DiktatRuleSet.kt
@@ -4,7 +4,6 @@ package org.cqfn.diktat.ruleset.rules
  * A group of [DiktatRule]'s as a single set.
  *
  * @property rules diktat rules.
- *
  */
 data class DiktatRuleSet(
     val rules: List<DiktatRule>
